# Halve runtime when simplifying large lines