--- a/lib/simplify.ex
+++ b/lib/simplify.ex
@@ -6,27 +6,34 @@ defmodule Simplify do
     simplify_dp_step(coordinates, tolerance * tolerance)
   end
 
-  @spec simplify(%Geo.LineString{}, number) :: %Geo.LineString{}
+  @spec simplify(Geo.LineString.t(), number) :: Geo.LineString.t()
   def simplify(%Geo.LineString{} = linestring, tolerance) do
     %Geo.LineString{coordinates: simplify(linestring.coordinates, tolerance)}
   end
 
-  defp simplify_dp_step(segment, _) when length(segment) < 3, do: segment
+  defp simplify_dp_step([], _), do: []
+  defp simplify_dp_step([_] = segment, _), do: segment
+  defp simplify_dp_step([_, _] = segment, _), do: segment
 
   defp simplify_dp_step(segment, tolerance_squared) do
-    first = List.first(segment)
-    last = List.last(segment)
+    [first | tail] = segment
+    {last, middle} = List.pop_at(tail, -1)
 
-    {far_index, _, far_squared_dist} =
-      Enum.zip(0..(length(segment) - 1), segment)
-      |> Enum.drop(1)
-      |> Enum.drop(-1)
-      |> Enum.map(fn {i, p} -> {i, p, seg_dist(p, first, last)} end)
-      |> Enum.max_by(&elem(&1, 2))
+    {_, far_value, far_index, far_squared_dist} =
+      Enum.reduce(middle, {1, nil, 1, 0}, fn element, {idx, max_val, max_idx, max_dist} ->
+        dist = seg_dist(element, first, last)
+
+        if dist >= max_dist do
+          {idx + 1, element, idx, dist}
+        else
+          {idx + 1, max_val, max_idx, max_dist}
+        end
+      end)
 
     if far_squared_dist > tolerance_squared do
-      front = simplify_dp_step(Enum.take(segment, far_index + 1), tolerance_squared)
-      [_ | back] = simplify_dp_step(Enum.drop(segment, far_index), tolerance_squared)
+      {pre_split, post_split} = Enum.split(segment, far_index + 1)
+      front = simplify_dp_step(pre_split, tolerance_squared)
+      [_ | back] = simplify_dp_step([far_value | post_split], tolerance_squared)
 
       front ++ back
     else