First, let me say thank you for putting together this library—it saved us a ton of time in our project. 🙏

This PR dramatically improves performance on lines with a large number of points, primarily by reducing the number of times we have to iterate over the list.

- Removes calls to `length/1`, since it's linear in the size of the list
- Cuts down on the number of times we access the last element of a list
- Cuts down on the number of new lists we construct
- Cuts down on the number of times we iterate the full list to find the maximum distance

This also adds a very simple benchmark to the test suite, which can be run with:

    $ mix test --only bench

Benchmarks on my machines (both running Elixir 13.4 + OTP 24, so the Intel machine has the benefit of the JIT):

| Machine | Per-iteration Runtime Before | After |
|--------|--------|------|
| 2019 Intel iMac  | ~0.15 sec | ~0.07 sec |
| 2020 M1 Macbook Air | ~0.19 sec | ~0.08 sec |