@pkinney 2022-05-28
Thanks, @s3cur3 !

I can definitely confirm a big performance boost (2022 M1 Max Macbook Pro):

Before:

```
## SimplifyBench
benchmark name                       iterations   average time
short linestring - high tolerance        200000   8.35 µs/op
short linestring - normal tolerance      100000   11.79 µs/op
short linestring - low tolerance         100000   12.00 µs/op
```

After: 

```
## SimplifyBench
benchmark name                       iterations   average time
short linestring - high tolerance       1000000   1.93 µs/op
short linestring - normal tolerance     1000000   2.68 µs/op
short linestring - low tolerance        1000000   2.85 µs/op
```

I'll merge this in then release as a 2.0.0 after I do a bit of long-overdue dependency updates and such.