@pkinney 2022-05-28
Released as part of [v2.0.0](https://hexdocs.pm/simplify/2.0.0)